requirements.txt
flask 
google-generativeai 
markdown 
bleach